move/rearrange extensions
